requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.0
dnspython==2.4.2
email-validator==2.0.0.post2
et-xmlfile==1.1.0
flake8==6.1.0
frozenlist==1.4.0
greenlet==3.0.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
mccabe==0.7.0
multidict==6.0.4
openpyxl==3.1.2
packaging==23.2
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
tabulate==0.9.0
typing_extensions==4.8.0
yarl==1.9.2
yoyo-migrations==8.2.0
zipp==3.17.0
